+++ components/org.wso2.carbon.identity.sso.saml/pom.xml
@@ -174,6 +174,10 @@
             </exclusions>
         </dependency>
         <dependency>
+            <groupId>org.wso2.carbon.identity.saml.common</groupId>
+            <artifactId>org.wso2.carbon.identity.saml.common.util</artifactId>
+        </dependency>
+        <dependency>
             <groupId>org.opensaml</groupId>
             <artifactId>opensaml-core</artifactId>
         </dependency>
@@ -325,6 +329,7 @@
                             org.wso2.carbon.identity.sp.metadata.saml2.exception; version="${identity.metadata.saml2.imp.pkg.version.range}",
                             org.wso2.carbon.identity.sp.metadata.saml2.util; version="${identity.metadata.saml2.imp.pkg.version.range}",
                             org.wso2.carbon.identity.core.model,
+                            org.wso2.carbon.identity.saml.common.util.*; version="${saml.common.util.version.range}",
                             org.wso2.carbon.idp.mgt.util,
                             javax.crypto,
                             javax.xml.stream,
